% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_adj.R
\name{seas_adjust}
\alias{seas_adjust}
\title{Tidy Seasonal Adjustment}
\usage{
seas_adjust(df, name_of_date_col = date, name_of_val_col = value)
}
\arguments{
\item{df}{A data frame containing dates and values as separate variables.}

\item{name_of_date_col}{Name of df column containing dates. Must be date formatted.}

\item{name_of_val_col}{Name of df column containing values to be seasonally adjusted.}
}
\description{
Wrapper for 'seasonal' package that plays more nicely with tidy data.
Input is a data frame (or similar object) that contains a date-formatted date variable and a numeric value variable.
Output is a vector of length equal to nrow(df) containing seasonall-adjusted values.
For now, output does *not* include any of the other outputs of the 'seas' function in the 'seasonal' package.
}
\examples{
df$SA <- function(df, date, NSA)
}
\keyword{seasonal}
