% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_adj.R
\name{seas_adjust}
\alias{seas_adjust}
\title{Tidy Seasonal Adjustment}
\usage{
seas_adjust(df, name_of_date_col = "date", name_of_val_col = "value")
}
\arguments{
\item{df}{A data frame}

\item{name_of_date_col}{Name of df column containing dates. Must be date formatted.}

\item{name_of_val_col}{Name of df column containing values. Must be numeric.}
}
\description{
Wrapper for 'seasonal' package.
}
\examples{
df$SA <- function(df, date, NSA)
}
\keyword{seasonal}
